Rally tasks added. File with tasks to perform is editable in templates/rally-tasks.
Some tests added to troubleshoot the instance (if not assigned to the compute1 host, try to reboot it) and
waiting image status to be active before booting the instance and waiting for instance to be active before pinging the instance.

Now try to implement the goto, based on checkpoints. Perform some tests first.